Rearranged components in MessageContainer and Messages such that Messages component is always rendered. Therefore we subscribe to newMessage subscription immediately upon login and despite selectedChat being null
